medallion_etl_pipeline/requirements.txt
# PySpark and Delta Lake
pyspark==3.5.0
delta-spark==3.0.0

# Prefect orchestration
prefect==2.14.21
prefect-dask==0.2.7

# Database connectivity (Python 3.13 compatible)
sqlalchemy==2.0.36  # Latest version with Python 3.13 support
psycopg[binary]==3.2.3  # Modern PostgreSQL driver with binary package
pandas==2.2.0
pyarrow>=15.0.0  # Required for pandas performance

# Excel processing
openpyxl==3.1.2
xlrd==2.0.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
numpy>=1.26.0